+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fix java 8 support, dropped due to jgit 6 transitive dependency (#13)
+
 ### 2.1.1 (2021-03-18)
 * Fix variables support ignore 'docs_dir' configuration from mkdocs.yml (#8)